send newsletter works, CRUD student works

+++ Admin_FE/src/app/help-tips/help-tips.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { HelpTip } from '../shared/help-tip';
 import { DataService } from '../services/data.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
+
 
 
 @Component({
@@ -25,7 +27,6 @@ export class HelpTipsComponent implements OnInit {
       let listHelpTips: any[] = result;
       listHelpTips.forEach( (element) => {
         this.arrHelpTips.push(element);
-        console.log(this.arrHelpTips)
       });
     });
   }
@@ -59,10 +60,26 @@ export class HelpTipsComponent implements OnInit {
       this.router.navigate(['/edit-helptip', help_ID]);
     }
     deleteHelpTip(help_ID: number){
-      this.dataService.DeleteHelpTip(help_ID).subscribe( (response:any) => {
-        window.location.reload();
+      Swal.fire({
+        title: 'Are you sure you want to delete this help tip?',
+        showConfirmButton: true,
+        showDenyButton: true,
+        confirmButtonText: 'Yes',
+        denyButtonText: 'No'
+      }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+          this.dataService.DeleteHelpTip(help_ID).subscribe( (response:any) => {
+            window.location.reload();
+          })
+          window.onload;    
+        } 
+        else if (result.isDenied) {
+          Swal.fire('Delete help tip has been aborted', '', 'warning')
+        }
       })
-      window.onload;
+
+
     }
 
 }
